Add basic support for self-referantial refs in copy mode
