+++ annotation-processor/src/main/kotlin/pl/touk/krush/source/RealReferencesMappingsGenerator.kt
@@ -1,34 +1,21 @@
 package pl.touk.krush.source
 
-import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FunSpec
-import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
-import com.squareup.kotlinpoet.asTypeName
-import org.jetbrains.exposed.sql.ResultRow
+import com.squareup.kotlinpoet.TypeName
 import pl.touk.krush.model.*
 import pl.touk.krush.model.AssociationType.*
 import pl.touk.krush.validation.EntityNotMappedException
-import pl.touk.krush.validation.MissingIdException
 import javax.lang.model.element.TypeElement
 
 class RealReferencesMappingsGenerator : MappingsGenerator() {
 
-    override fun buildToEntityMapFunc(entityType: TypeElement, entity: EntityDefinition, graphs: EntityGraphs): FunSpec {
-        val rootKey = entity.id?.asTypeName() ?: throw MissingIdException(entity)
-
-        val rootVal = entity.name.asVariable()
-        val func = FunSpec.builder("to${entity.name}Map")
-                .receiver(Iterable::class.parameterizedBy(ResultRow::class))
-                .returns(ClassName("kotlin.collections", "MutableMap").parameterizedBy(rootKey, entityType.asType().asTypeName()))
-
-        val rootIdName = entity.id.name.asVariable()
-        val rootValId = "${rootVal}Id"
-
-        func.addStatement("var roots = mutableMapOf<$rootKey, ${entity.name}>()")
+    override fun buildToEntityMapFuncBody(entityType: TypeElement, entity: EntityDefinition, graphs: EntityGraphs, func: FunSpec.Builder,
+                                          entityId: IdDefinition, rootKey: TypeName, rootVal: String, rootIdName: String, rootValId: String): FunSpec {
+        func.addStatement("val roots = mutableMapOf<$rootKey, ${entity.name}>()")
 
         // Add all non-relational data
         func.addStatement("this.forEach { resultRow ->")
-        func.addStatement("\tval $rootValId = resultRow.getOrNull(${entity.name}Table.${entity.id.name}) ?: return@forEach")
+        func.addStatement("\tval $rootValId = resultRow.getOrNull(${entity.name}Table.${entityId.name}) ?: return@forEach")
         func.addStatement("\tval $rootVal = roots[$rootValId] ?: resultRow.to${entity.name}()")
         func.addStatement("\troots[$rootValId] = $rootVal")
         func.addStatement("}")
@@ -36,11 +23,11 @@ class RealReferencesMappingsGenerator : MappingsGenerator() {
         val associations = entity.getAssociations(ONE_TO_ONE, ONE_TO_MANY, MANY_TO_MANY)
         associations.forEach { assoc ->
             val target = graphs[assoc.target.packageName]?.get(assoc.target) ?: throw EntityNotMappedException(assoc.target)
-            val entityTypeName = entity.id.asTypeName()
+            val entityIdTypeName = entityId.asTypeName()
             val associationMapName = "${entity.name.asVariable()}_${assoc.name}"
             val associationMapValueType = if (assoc.type in listOf(ONE_TO_MANY, MANY_TO_MANY)) "MutableSet<${target.name}>" else "${target.name}"
 
-            func.addStatement("val $associationMapName = mutableMapOf<${entityTypeName}, $associationMapValueType>()")
+            func.addStatement("val $associationMapName = mutableMapOf<${entityIdTypeName}, $associationMapValueType>()")
             if (!(assoc.type == ONE_TO_ONE && assoc.mapped)) {
                 val isSelfReferential = assoc.target == entityType
 
@@ -49,14 +36,13 @@ class RealReferencesMappingsGenerator : MappingsGenerator() {
                 if (!isSelfReferential) {
                     func.addStatement("val ${assoc.name}_map = this.to${assoc.target.simpleName}Map()")
                 }
-
             }
         }
 
         // Add O2O relational data first, because it recreates the entity objects which destroys potential object references
         if (associations.any { it.type == ONE_TO_ONE }) {
             func.addStatement("this.forEach { resultRow ->")
-            func.addStatement("\tval $rootValId = resultRow.getOrNull(${entity.name}Table.${entity.id.name}) ?: return@forEach")
+            func.addStatement("\tval $rootValId = resultRow.getOrNull(${entity.name}Table.${entityId.name}) ?: return@forEach")
             func.addStatement("\tval $rootVal = roots[$rootValId] ?: resultRow.to${entity.name}()")
             associations.forEach { assoc ->
                 if (assoc.type != ONE_TO_ONE) {
@@ -100,7 +86,7 @@ class RealReferencesMappingsGenerator : MappingsGenerator() {
 
             // Add list relational data (this is done in a separate step so that self-referential relations work)
             func.addStatement("this.forEach { resultRow ->")
-            func.addStatement("\tval $rootValId = resultRow.getOrNull(${entity.name}Table.${entity.id.name}) ?: return@forEach")
+            func.addStatement("\tval $rootValId = resultRow.getOrNull(${entity.name}Table.${entityId.name}) ?: return@forEach")
             func.addStatement("\tval $rootVal = roots[$rootValId] ?: resultRow.to${entity.name}()")
             associations.forEach { assoc ->
                 val target = graphs[assoc.target.packageName]?.get(assoc.target) ?: throw EntityNotMappedException(assoc.target)
