+++ example/src/main/kotlin/pl/touk/krush/many2many/Task.kt
@@ -2,8 +2,7 @@ package pl.touk.krush.many2many
 
 import javax.persistence.*
 
-// Uncomment to test M2M with real references (requires krush.references to be set to real)
-/*@Entity
+@Entity
 @Table(name = "tasks")
 data class Task(
 
@@ -15,4 +14,4 @@ data class Task(
         @JoinTable(name = "task_requirements")
         @ManyToMany
         val requirements: Collection<Task> = emptyList()
-)*/
+)
