+++ example/src/test/kotlin/pl/touk/krush/many2many/TaskTest.kt
@@ -6,13 +6,14 @@ import org.jetbrains.exposed.sql.JoinType
 import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.transaction
+import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import pl.touk.krush.base.BaseDatabaseTest
 
-// Uncomment to test M2M with real references (requires krush.references to be set to real)
-/*class TaskTest : BaseDatabaseTest() {
+class TaskTest : BaseDatabaseTest() {
 
     @Test
+    @Disabled("Uncomment to test M2M with real references (requires krush.references to be set to real)")
     fun shouldHandleSelfReferences() {
         transaction {
             SchemaUtils.create(TaskTable, TaskRequirementsTable)
@@ -38,4 +39,4 @@ import pl.touk.krush.base.BaseDatabaseTest
         }
     }
 
-}*/
+}
